move checkpoint sync code to a common location
